requirements.txt
Flask==2.3.2
Flask-JWT-Extended==4.5.2
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0